Increase priority of KLBFormat
SCIFIO probes the format of a file sequentially. The order in which formats are tried is influenced
by the 'priority' annotation. Massively increasing its value improves KLB reading performance,
because it is now tried before most other formats (some of which require reading from the file).
This also prevents mis-detection errors (as encountered recently), and won't decrease the
performance of other formats because detecting KLB format is cheap (file name extension is
necessary and sufficient).